requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Werkzeug
psycopg2-binary
gunicorn
flask-login
bcrypt
cerberus
bleach
flask-babel
deep-translator
Filerobot 
minio

